refactor: split engine logic
